requirements.txt
discord.py
sqlalchemy
asyncio